+++ CHANGELOG.md
@@ -8,17 +8,18 @@
 * Added a `onCancel` parameter to the `Menu` constructor, so the `cancel` method doesn't need to be overridden every time.
 * When creating a `Menu` instance, it is now possible to set the initial position within the menu.
 * You can now specify random sounds for `Menu` instances.
+* It is now possible to specify multiple triggers with the same name.
 
 ### Fixed
 
 * Fixed a broken link in the documentation for `angleToRad`.
 
-### Changes
+### Changed
 
 * Cleaned up code.
 * Update SDL.
 * Added a bunch of parameters to the `Menu` constructor, to replace the old `registerCommands` method.
-* It is now an error to register multiple commands with the same name.
+* The `TriggerMap` class now uses a list for command triggers.
 
 ### Removed